后台管理功能

+++ merchant-loans-admin/src/main/java/cn/com/payu/modules/loans/req/ApplymentRejectReq.java
@@ -11,7 +11,7 @@ public class ApplymentRejectReq extends AbstractApiReq {
     @SerializedName("order_number")
     private String orderNumber;//是	string	32	我方订单号
     private String reason;//是	string	200	驳回原因
-    private BigDecimal annuity;//是	int	10	申请金额
+    private Integer annuity;//是	int	10	申请金额
     private Integer loanPeriod;//是	int	10	申请期限
 
 }
